Decompose MoneyRequestView into data hook and field components
